+++ example/lib/main.dart
@@ -11,7 +11,7 @@ class ImageClipExampleApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: '图片剪辑',
+      title: 'Image Clip',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
         useMaterial3: true,